+++ src/ballistica/core/core.cc
@@ -123,10 +123,10 @@ void CoreFeatureSet::PostInit() {
   // We can use it to make error messages/etc. more pretty by stripping out
   // all but sub-project paths.
   const char* f = __FILE__;
-  auto* f_end = strstr(
-      f, "src" BA_DIRSLASH "ballistica" BA_DIRSLASH "app" BA_DIRSLASH "app.cc");
-  if (!f) {
-    Log(LogLevel::kWarning, "Unable to calc project dir from __FILE__.");
+  auto* f_end = strstr(f, "src" BA_DIRSLASH "ballistica" BA_DIRSLASH
+                          "core" BA_DIRSLASH "core.cc");
+  if (!f_end) {
+    Log(LogLevel::kWarning, "Unable to calc build source dir from __FILE__.");
   } else {
     build_src_dir_ = std::string(f).substr(0, f_end - f);
   }